+++ src/pytest_inmanta_lsm/remote_orchestrator.py
@@ -8,8 +8,8 @@
 
 import logging
 import os
-import shlex
 import subprocess
+from pathlib import Path
 from pprint import pformat
 from typing import Dict, Optional, Union
 from uuid import UUID
@@ -105,6 +105,7 @@ class RemoteOrchestrator:
         self._server_cache_path: Optional[str] = None
 
         self._ensure_environment()
+        self._server_version = self._get_server_version()
 
     @property
     def environment(self) -> UUID:
@@ -120,6 +121,25 @@ class RemoteOrchestrator:
     @property
     def host(self) -> str:
         return self._host
+
+    @property
+    def server_version(self) -> Version:
+        """
+        Returns the version of the remote orchestrator
+        """
+        return self._server_version
+
+    def _get_server_version(self) -> Version:
+        """
+        Get the version of the remote orchestrator
+        """
+        server_status: Result = self.client.get_server_status()
+        if server_status.code != 200:
+            raise Exception(f"Failed to get server status for {self._host}")
+        try:
+            return Version(server_status.result["data"]["version"])
+        except (KeyError, TypeError):
+            raise Exception(f"Unexpected response for server status API call: {server_status.result}")
 
     def export_service_entities(self) -> None:
         """Initialize the remote orchestrator with the service model and check if all preconditions hold"""
@@ -169,6 +189,11 @@ class RemoteOrchestrator:
     def sync_project(self) -> None:
         """Synchronize the project to the lab orchestrator"""
         project = self._project
+
+        source_script = Path(__file__).parent / "resources/setup_project.py"
+        destination_script = Path(project._test_project_dir, ".inm_lsm_setup_project.py")
+        LOGGER.debug(f"Copying module V2 install script ({source_script}) in project folder {destination_script}")
+        destination_script.write_text(source_script.read_text())
 
         LOGGER.info("Sending service model to the lab orchestrator")
         # load the project yaml
@@ -266,33 +291,25 @@ class RemoteOrchestrator:
             stderr=subprocess.PIPE,
         )
 
-        server_status: Result = self.client.get_server_status()
-        if server_status.code != 200:
-            raise Exception(f"Failed to get server status for {self._host}")
-        server_version: Version
-        try:
-            server_version = Version(server_status.result["data"]["version"])
-        except (KeyError, TypeError):
-            raise Exception("Unexpected response for server status API call")
         # iso5 requires explicit project installation
-        if server_version >= Version("5.dev"):
-            venv_path: str = os.path.join(server_path, ".env")
+        if self.server_version >= Version("5.dev"):
+            LOGGER.debug(f"Server version is {self.server_version}, installing project manually")
             # venv might not exist yet so can't just access its `inmanta` executable -> install via Python script instead
-            python_script_inline: str = (
-                "from inmanta.module import Project;"
-                f"project = Project('{server_path}', venv_path='{venv_path}');"
-                "project.install_modules();"
-            )
-            shell_script_inline: str = "/opt/inmanta/bin/python -c %s" % shlex.quote(python_script_inline)
+            install_script_path = Path(server_path, destination_script.name)
+            shell_script_inline: str = f"/opt/inmanta/bin/python < {install_script_path}"
             if not self.container_env:
                 # use the server's environment variables for the installation
                 shell_script_inline = (
                     f"{use_sudo}systemd-run --pipe -p User=inmanta -p EnvironmentFile=/etc/sysconfig/inmanta-server "
+                    f"-p Environment=PROJECT_PATH={server_path} "
                     "--wait %s" % shell_script_inline
                 )
+            else:
+                # Add the project path as env var at the beginning of the cmd line
+                shell_script_inline = f"PROJECT_PATH={server_path} {shell_script_inline}"
 
             try:
-                subprocess.check_output(
+                output = subprocess.check_output(
                     SSH_CMD
                     + [
                         f"-p {self._ssh_port}",
@@ -300,7 +317,10 @@ class RemoteOrchestrator:
                         shell_script_inline,
                     ],
                     stderr=subprocess.PIPE,
+                    encoding="utf-8",
+                    text=True,
                 )
+                LOGGER.debug(output)
             except subprocess.CalledProcessError as e:
                 LOGGER.error("Process failed out: " + e.output.decode())
                 LOGGER.error("Process failed err: " + e.stderr.decode())